+++ src/components/UserCard.tsx
@@ -1,21 +1,27 @@
-import React, { FC } from 'react';
+import React, { FC, memo } from 'react';
 import NextLink from 'next/link';
-import { Avatar, Link, Text, WrapItem } from '@chakra-ui/react';
-import type { GitHubUser } from '@/server/user';
+import { Avatar, Link, Text, VStack, WrapItem } from '@chakra-ui/react';
+import UserFollowLabel from './UserFollowLabel';
 
-type UserCardProps = GitHubUser;
+interface UserCardProps {
+  avatarUrl: string;
+  username: string;
+}
 
 const UserCard: FC<UserCardProps> = ({ avatarUrl, username }) => (
   <WrapItem m={[2, 4]}>
-    <NextLink href={`/user/${username}`} passHref>
-      <Link textAlign="center" maxW={[28, 36, 44]}>
-        <Avatar name={username} src={avatarUrl} size="xl" />
-        <Text mx="2" isTruncated>
-          {username}
-        </Text>
-      </Link>
-    </NextLink>
+    <VStack>
+      <NextLink href={`/user/${username}`} passHref>
+        <Link textAlign="center" maxW={[28, 36, 44]}>
+          <Avatar name={username} src={avatarUrl} size="xl" />
+          <Text mx="2" isTruncated>
+            {username}
+          </Text>
+        </Link>
+      </NextLink>
+      <UserFollowLabel username={username} />
+    </VStack>
   </WrapItem>
 );
 
-export default UserCard;
+export default memo(UserCard);