+++ src/pages/api/follow/[username].ts
@@ -1,18 +1,22 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { getUserByUsername } from '@/server/user';
-import type { GetUserResponse } from '@/redux/api';
+import { getUserFollowersAndFollowing } from '@/server/user';
+import { GetUserFollowCount } from '@/redux/api';
 
 export default async (
   req: NextApiRequest,
-  res: NextApiResponse<GetUserResponse>,
+  res: NextApiResponse<GetUserFollowCount>,
 ): Promise<void> => {
   const username = req.query.username as string;
 
   switch (req.method) {
     case 'GET': {
       try {
-        const user = await getUserByUsername(username);
-        return res.status(200).json({ success: true, user });
+        const userFollow = await getUserFollowersAndFollowing(username);
+        return res.status(200).json({
+          success: true,
+          followerCount: userFollow.followers.length,
+          followingCount: userFollow.followings.length,
+        });
       } catch (error) {
         console.error(error);
         return res.status(404).json({ success: false });