+++ src/pages/api/users/index.ts
@@ -1,9 +1,10 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { searchUsersByUsername } from '@/server/user';
+import { searchUsersByUsername } from '@/server/search-user';
+import type { SearchUserResponse } from '@/redux/api';
 
 export default async (
   req: NextApiRequest,
-  res: NextApiResponse,
+  res: NextApiResponse<SearchUserResponse>,
 ): Promise<void> => {
   switch (req.method) {
     case 'GET': {
@@ -14,14 +15,18 @@ export default async (
         return res.status(400).json({ success: false });
       }
 
-      const { isEnd, total, users } = await searchUsersByUsername(
-        username,
-        parseInt(page, 10),
-      );
-
-      return res
-        .status(200)
-        .json({ success: users.length > 0, isEnd, total, users });
+      try {
+        const result = await searchUsersByUsername(
+          username,
+          parseInt(page, 10),
+        );
+        return res
+          .status(200)
+          .json({ success: result.users.length > 0, ...result });
+      } catch (error) {
+        console.error(error);
+        return res.status(400).json({ success: false });
+      }
     }
 
     default: