+++ src/pages/user/[username].tsx
@@ -10,42 +10,25 @@ import {
   Heading,
   Box,
 } from '@chakra-ui/react';
-import type { GetServerSideProps } from 'next';
-import {
-  SingleGitHubUser,
-  getUserByUsername,
-  getFollowersByUsername,
-  getFollowingsByUsername,
-  getRepositoriesByUsername,
-} from '@/server/user';
+import type { GetStaticPaths, GetStaticProps } from 'next';
+import { getUserByUsername, User } from '@/server/user';
 
-interface UserPageProps {
-  user: SingleGitHubUser;
-  repositories: string[];
-  followers: string[];
-  followings: string[];
-}
+type UserPageProps = User;
+
+const FIVE_MINUTES = 5 * 60;
+
+const renderName = (name: string) => <Text key={name}>{name}</Text>;
 
 const UserPage: FC<UserPageProps> = ({
-  user,
+  info,
   repositories,
   followers,
   followings,
 }) => {
   const { isOpen, onToggle } = useDisclosure();
 
-  if (!user) {
-    return (
-      <Container as="main" p={[4, 8, 12]} minH="80vh" centerContent>
-        <Text>Not Found</Text>
-      </Container>
-    );
-  }
-
-  const renderName = (name: string) => <Text key={name}>{name}</Text>;
-
   // FIXME: 'plan' exists when accessing tokens' owner, should remove in server
-  const { plan, ...rest } = user;
+  const { plan, ...rest } = info;
 
   return (
     <Container as="main" p={[4, 8, 12]} minH="80vh" centerContent>
@@ -68,7 +51,7 @@ const UserPage: FC<UserPageProps> = ({
             {Object.keys(rest).map((key) => (
               <HStack key={key} alignItems="start">
                 <Text fontWeight="bold">{key}</Text>
-                <Text>{user[key]}</Text>
+                <Text>{info[key]}</Text>
               </HStack>
             ))}
           </VStack>
@@ -78,22 +61,22 @@ const UserPage: FC<UserPageProps> = ({
   );
 };
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
+export const getStaticPaths: GetStaticPaths = async () => ({
+  paths: [],
+  // blocking will server-render new page and cache
+  fallback: 'blocking',
+});
+
+export const getStaticProps: GetStaticProps = async (context) => {
   const username = context.params.username as string;
-  const user = await getUserByUsername(username);
-  let repositories: string[];
-  let followers: string[];
-  let followings: string[];
 
-  if (user !== null) {
-    [repositories, followers, followings] = await Promise.all([
-      getRepositoriesByUsername(username),
-      getFollowersByUsername(username),
-      getFollowingsByUsername(username),
-    ]);
+  try {
+    const user = await getUserByUsername(username);
+    return { props: user, revalidate: FIVE_MINUTES };
+  } catch (error) {
+    console.error(error);
+    return { notFound: true };
   }
-
-  return { props: { user, repositories, followers, followings } };
 };
 
 export default UserPage;