+++ src/redux/user/slice.ts
@@ -1,10 +1,10 @@
 import { createSlice } from '@reduxjs/toolkit';
-import type { GitHubUser } from '@/server/user';
+import type { UserProfile } from '@/server/user';
 import { search } from './action';
 
 export interface UserState {
   isSearching: boolean;
-  users: GitHubUser[];
+  users: UserProfile[];
   message?: string;
 }
 
@@ -28,7 +28,7 @@ const { reducer } = createSlice({
     });
     builder.addCase(search.rejected, (state, action) => {
       state.isSearching = false;
-      state.message = `Cannot find ${action.meta.arg}, try another one?`;
+      state.message = `Cannot find ${action.meta.arg.username}, try another one?`;
     });
   },
 });